requirements.txt
py-cord 
python-dotenv
pymongo
httpx
